Refactor: 사물함 색상 함수 분리 + 검색 결과에 따른 사물함 API 호출 중복함수 제거#50
