+++ src/api/searchKeywordApi.tsx
@@ -1,11 +1,6 @@
-// Search API 호출
-
-import axios from "axios";
-const SEARCH_URL = import.meta.env.VITE_SEARCH_URL;
-
 import api from "@/api/axiosInterceptApi";
 
-// // Search API 호출
+// Search API 호출
 export const searchKeywordApi = async (keyword) => {
   try {
     const response = await api.get(`/cabinet/search?keyword=${keyword}`);
