+++ src/components/Cabinet/CabinetButtonLayout.tsx
@@ -1,17 +1,17 @@
 // 사물함 배열 관련
 
+import { useEffect } from "react";
 import { cabinetDetailInfoApi } from "@/api/cabinetDetailInfoApi";
 import { useCabinetData } from "@/hooks/useCabinetData";
-import { useEffect } from "react";
+import { useCabinetState } from "@/hooks/useCabinetState";
 
 interface CabinetButtonLayoutProps {
-  selectedBuilding: { name: string; floors: number } | null;
+  selectedBuilding: { name: string } | null;
   selectedFloor: number | null;
   selectedCabinet: { cabinetId: number; cabinetNumber: number } | null;
   setSelectedCabinet: (
     cabinet: { cabinetId: number; cabinetNumber: number } | null
   ) => void;
-  selectedStatus: string;
   setSelectedStatus: (status: string) => void;
   isMyCabinet: boolean;
   setIsMyCabinet: (isMine: boolean) => void;
@@ -28,7 +28,6 @@ const CabinetButtonLayout = ({
   selectedFloor,
   selectedCabinet,
   setSelectedCabinet,
-  selectedStatus,
   setSelectedStatus,
   isMyCabinet,
   setIsMyCabinet,
@@ -38,12 +37,13 @@ const CabinetButtonLayout = ({
     selectedBuilding,
     selectedFloor,
     selectedCabinet,
-    isMyCabinet,
-    selectedStatus
+    isMyCabinet
   );
 
+  const { getStatusColor } = useCabinetState();
+
   // 사물함 정보 API 호출
-  const fetchCabinetDetailInformaion = async (
+  const fetchCabinetDetailInformation = async (
     cabinetId: number,
     cabinetNumber: number
   ) => {
@@ -58,6 +58,7 @@ const CabinetButtonLayout = ({
       console.error(error);
     }
   };
+
   useEffect(() => {
     if (filteredCabinetDetail) {
       // cabinetNumber로 cabinetData에서 매칭되는 cabinet 찾기
@@ -68,7 +69,7 @@ const CabinetButtonLayout = ({
 
       if (matchedCabinet) {
         // cabinetId를 사용하여 상세 정보 API 호출
-        fetchCabinetDetailInformaion(
+        fetchCabinetDetailInformation(
           matchedCabinet.id,
           matchedCabinet.cabinetNumber
         );
@@ -76,26 +77,6 @@ const CabinetButtonLayout = ({
     }
     return;
   }, [cabinetData]);
-
-  // 각 상태에 대한 버튼 색상 설정
-  const getStatusColor = (selectedStatus: string, isMyCabinet: boolean) => {
-    if (selectedStatus === "USING") {
-      if (isMyCabinet === true) {
-        return "bg-lime-500 text-white"; // 본인이 사용 중인 사물함
-      }
-      if (isMyCabinet === false) {
-        return "bg-purple-500 text-white"; // 다른 사람이 사용 중인 사물함
-      }
-    }
-    switch (selectedStatus) {
-      case "OVERDUE":
-        return "bg-red-500 text-white"; // 반납 지연
-      case "AVAILABLE":
-        return "bg-gray-300"; // 이용 가능
-      case "BROKEN":
-        return "bg-gray-700 text-white"; // 사용 불가
-    }
-  };
 
   return (
     <div className="w-full h-[80%] flex items-center justify-center">
@@ -113,7 +94,10 @@ const CabinetButtonLayout = ({
                 left: `${cabinet.cabinetXPos * 90}px`, // API에서 받은 xPos 사용
               }}
               onClick={() => {
-                fetchCabinetDetailInformaion(cabinet.id, cabinet.cabinetNumber);
+                fetchCabinetDetailInformation(
+                  cabinet.id,
+                  cabinet.cabinetNumber
+                );
               }}
             >
               {cabinet.cabinetNumber}
