+++ src/components/Cabinet/Modal/CabinetRentalConfirmModal.tsx
@@ -57,7 +57,9 @@ const CabinetRentalConfirmModal = ({
         <div className="mt-5 flex justify-center">
           <button
             className="mr-2 px-4 py-2 bg-blue-600 text-white rounded-md hover:bg-blue-500"
-            onClick={fetchCabinetRental} // 확인 버튼 클릭 시 대여 API 호출 + 모달 닫음 + cabinetRentalComplete.tsx가 렌더링
+            onClick={fetchCabinetRental}
+
+            // 확인 버튼 클릭 시 대여 API 호출 + 모달 닫음 + cabinetRentalComplete.tsx가 렌더링
           >
             확인
           </button>
