+++ src/hooks/useCabinetState.tsx
@@ -1,4 +1,4 @@
-// 사물함 선택, 대여, 반납 관련 hook
+// 사물함 선택, 대여, 반납, color 관련 hook
 
 import { cabinetDetailInfoApi } from "@/api/cabinetDetailInfoApi";
 import { useEffect, useState } from "react";
@@ -28,6 +28,25 @@ export const useCabinetState = () => {
     }
   };
 
+  // 각 상태에 대한 버튼 색상 설정
+  const getStatusColor = (selectedStatus: string, isMyCabinet: boolean) => {
+    if (selectedStatus === "USING") {
+      if (isMyCabinet === true) {
+        return "bg-lime-500 text-white"; // 본인이 사용 중인 사물함
+      } else {
+        return "bg-purple-500 text-white"; // 다른 사람이 사용 중인 사물함
+      }
+    }
+    switch (selectedStatus) {
+      case "OVERDUE":
+        return "bg-red-500 text-white"; // 반납 지연
+      case "AVAILABLE":
+        return "bg-gray-300"; // 이용 가능
+      case "BROKEN":
+        return "bg-gray-700 text-white"; // 사용 불가
+    }
+  };
+
   useEffect(() => {
     if (selectedCabinet === null) return;
     if (selectedCabinet.cabinetId) {
@@ -46,5 +65,6 @@ export const useCabinetState = () => {
     isMyCabinet,
     setIsMyCabinet,
     fetchCabinetDetailInformation,
+    getStatusColor,
   };
 };
